fix of OSXDriver cos of CFIndex being long instead of int returning negative big values instead of -32768 added rounding of axis values with dread zone up/down to 0 and -1/1 added code for handling triggers to give value from 0 to 1

+++ Assets/Scripts/ws/winx/platform/OSX/Native.cs
@@ -49,7 +49,7 @@ namespace ws.winx.platform.osx
 	//using IOReturn = //System.IntPtr;
 	using IOHIDElementCookie = System.UInt32;
 	using CFTypeID=System.Int32;//System.UInt64;
-	using CFIndex =System.Int64;
+	using CFIndex =System.Int32;
 	using CFTimeInterval=System.Double;
 	using mach_port_t=System.UInt32;
 	using UInt8=System.Byte;
